requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
Django==2.2.12
django-bootstrap-pagination==1.7.1
django-bootstrap-ui==1.1.0
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-project==0.7
django-taggit==1.2.0
Markdown==3.2.1
Pillow==7.0.0
pytz==2019.3
sorl-thumbnail==12.6.3
soupsieve==2.0
sqlparse==0.3.1
